pictures upload works on route /imageUpload
