+++ src/App.js
@@ -5,6 +5,7 @@ import User from './user/User';
 import Admin from './admin/Admin'
 import CreatePost from './user/landing/create-post/CreatePost';
 import Explore from './admin/landing/explore/Explore';
+import ImageUpload from './ImageUpload/ImageUpload';
 class App extends Component {
  render() {
    return (
@@ -17,6 +18,7 @@ class App extends Component {
          <Route path="/admin" exact render={() => <Admin />}/>
          <Route path="/admin/explore" exact render={() => <Explore />} />
           {/* <Route path="/admin/search" exact render={() => <Search />} /> */}
+          <Route path="/imageUpload" exact render={() => <ImageUpload/>}/>
        </div>
      </Router>
    );
