+++ src/Firebase/index.js
@@ -14,4 +14,9 @@ const firebaseConfig = {
 };
 // Initialize Firebase
 firebase.initializeApp(firebaseConfig);
-firebase.analytics();
+
+const storage = firebase.storage()
+
+export {
+    storage, firebase as default
+}
